+++ test/codegen/joinpoint_mixed_arg_types.mlir
@@ -1,17 +1,17 @@
 // RUN: %ecoc %s -emit=jit 2>&1 | %FileCheck %s
 //
-// Test joinpoint with multiple argument types: i64, f64, i32, and !eco.value.
+// Test joinpoint with multiple argument types: i64, f64, i16, and !eco.value.
 // This tests type conversion handling in joinpoint argument passing.
 
 module {
   func.func @main() -> i64 {
     %int_val = arith.constant 42 : i64
     %float_val = arith.constant 3.14 : f64
-    %char_val = arith.constant 65 : i32  // 'A'
+    %char_val = arith.constant 65 : i16  // 'A'
     %boxed = eco.box %int_val : i64 -> !eco.value
 
     // Joinpoint with mixed argument types
-    eco.joinpoint 0(%i: i64, %f: f64, %c: i32, %v: !eco.value) {
+    eco.joinpoint 0(%i: i64, %f: f64, %c: i16, %v: !eco.value) {
       // Use all the arguments
       eco.dbg %i : i64
       // CHECK: 42
@@ -19,7 +19,7 @@ module {
       eco.dbg %f : f64
       // CHECK: 3.14
 
-      eco.dbg %c : i32
+      eco.dbg %c : i16
       // CHECK: 'A'
 
       eco.dbg %v : !eco.value
@@ -27,7 +27,7 @@ module {
 
       eco.return
     } continuation {
-      eco.jump 0(%int_val, %float_val, %char_val, %boxed : i64, f64, i32, !eco.value)
+      eco.jump 0(%int_val, %float_val, %char_val, %boxed : i64, f64, i16, !eco.value)
     }
 
     // Test 2: Joinpoint with f64 argument
